+++ internal/domain/repositories/task_repository.go
@@ -12,7 +12,7 @@ var ErrTaskNotFound = errors.New("taqque-internal-domain-repositories-task-repos
 type TaskRepository interface {
 	GetMultiOrderByPriorityDescAndSortKeyAsc(ctx context.Context, limit *int) ([]*model.Task, error)
 	Create(ctx context.Context, task *model.Task) error
-	GetFirstByPriorityOrderBySortKeyAsc(ctx context.Context, priority int) (*model.Task, error)
+	GetFirstByPriorityOrderBySortKeyAsc(ctx context.Context, priority float64) (*model.Task, error)
 	Update(ctx context.Context, task *model.Task) error
 	// UpdateMultiAsDeleted(ctx context.Context, id string) error
 }
